+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2Configuration.scala
@@ -7,6 +7,7 @@ import pl.touk.nussknacker.engine.util.config.FicusReaders.forDecoder
 import pl.touk.nussknacker.ui.security.CertificatesAndKeys
 import pl.touk.nussknacker.ui.security.api.{AuthenticationConfiguration, FrontendStrategySettings}
 import pl.touk.nussknacker.ui.security.oauth2.ProfileFormat.ProfileFormat
+import pl.touk.nussknacker.ui.security.oauth2.UsernameFieldName.UsernameFieldName
 import sttp.model.{HeaderNames, MediaType, Uri}
 
 import java.net.URI
@@ -35,6 +36,7 @@ case class OAuth2Configuration(usersFile: URI,
                                anonymousUserRole: Option[String] = None,
                                tokenCookie: Option[TokenCookieConfig] = None,
                                overrideFrontendAuthenticationStrategy: Option[FrontendStrategySettings] = None,
+                               usernameFieldName: Option[UsernameFieldName] = None,
                               ) extends AuthenticationConfiguration {
   override def name: String = OAuth2Configuration.name
 
@@ -70,6 +72,14 @@ object ProfileFormat extends Enumeration {
   type ProfileFormat = Value
   val GITHUB = Value("github")
   val OIDC = Value("oidc")
+}
+
+object UsernameFieldName extends Enumeration {
+  type UsernameFieldName = Value
+  val PreferredUsername: UsernameFieldName = Value("preferred_username")
+  val GivenName: UsernameFieldName = Value("given_name")
+  val Nickname: UsernameFieldName = Value("nickname")
+  val Name: UsernameFieldName = Value("name")
 }
 
 trait JwtConfiguration {