+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oidc/OidcAuthenticationConfiguration.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.ui.security.oidc
 
 import pl.touk.nussknacker.engine.util.config.URIExtensions
 import pl.touk.nussknacker.ui.security.oauth2.ProfileFormat.OIDC
+import pl.touk.nussknacker.ui.security.oauth2.UsernameFieldName.UsernameFieldName
 import pl.touk.nussknacker.ui.security.oauth2.{JwtConfiguration, OAuth2Configuration, TokenCookieConfig}
 import sttp.client3.SttpBackend
 import sttp.model.MediaType
@@ -29,6 +30,7 @@ case class OidcAuthenticationConfiguration(usersFile: URI,
                                            jwksUri: Option[URI] = None,
                                            rolesClaims: Option[List[String]] = None,
                                            tokenCookie: Option[TokenCookieConfig] = None,
+                                           usernameFieldName: Option[UsernameFieldName] = None,
                                            accessTokenIsJwt: Boolean = false,
                                           ) extends URIExtensions {
 
@@ -61,7 +63,8 @@ case class OidcAuthenticationConfiguration(usersFile: URI,
     accessTokenParams = Map("grant_type" -> "authorization_code"),
     accessTokenRequestContentType = MediaType.ApplicationXWwwFormUrlencoded.toString(),
     anonymousUserRole = anonymousUserRole,
-    tokenCookie = tokenCookie
+    tokenCookie = tokenCookie,
+    usernameFieldName = usernameFieldName
   )
 
   def withDiscovery(implicit ec: ExecutionContext, sttpBackend: SttpBackend[Future, Any]): OidcAuthenticationConfiguration = {